Refactor turbo and baxus to use the same interface
