+++ bbopy/algorithms/botorch_algorithms/baxus.py
@@ -1,6 +1,6 @@
 import math
 from dataclasses import dataclass
-from typing import Union, Type, Optional, Dict, Any, Tuple
+from typing import Union, Type, Optional, Dict, Any
 
 import gpytorch
 import numpy as np
@@ -9,21 +9,18 @@ from botorch import fit_gpytorch_mll
 from botorch.acquisition import ExpectedImprovement
 from botorch.exceptions import ModelFittingError
 from botorch.generation import MaxPosteriorSampling
-from botorch.models import SingleTaskGP
-from botorch.optim import optimize_acqf
 from gpytorch import ExactMarginalLogLikelihood
 from gpytorch.constraints import Interval
 from gpytorch.kernels import ScaleKernel, MaternKernel
 from gpytorch.likelihoods import GaussianLikelihood
 from torch import Tensor
-from torch.quasirandom import SobolEngine
 
-from bbopy.algorithms import BoTorchAlgorithm
+from bbopy.algorithms.botorch_algorithms import TURBO
 from bbopy.problems.base import Problem
 from bbopy.sampling import Sampling, FloatRandomSampling
 
 
-class BAxUS(BoTorchAlgorithm):
+class BAxUS(TURBO):
     name: str = "BAxUS"
 
     def __init__(
@@ -34,27 +31,19 @@ class BAxUS(BoTorchAlgorithm):
             acquisition_optimizer_kwargs: Optional[Dict[str, Any]] = None,
             sampling: Sampling = FloatRandomSampling(backend="torch", dtype=torch.double),
     ) -> None:
-        super().__init__(n_init)
+        super().__init__(n_init, acquisition, acquisition_optimizer_kwargs, sampling)
         self._target_bounds = None
+        self._train_x_target = None
         self.evaluation_budget = evaluation_budget
         self.S = None
-        self._dim = None
-        self.state = None
-        self.n_candidates = None
-        self._sampling = sampling
-        self._surrogate = SingleTaskGP
-        self._acquisition = acquisition
-        self._acquisition_optimizer_kwargs = acquisition_optimizer_kwargs
-        if acquisition_optimizer_kwargs is None:
-            _, self._acquisition_optimizer_kwargs = self._default_acqf_optimizer()
-
-        self._train_x_target = None
+        self.batch_size = 1
 
     def setup(
             self,
             problem: Problem,
     ) -> None:
         self._dim = problem.dim
+        self._acqf_optimizer_kwargs = self._acqf_optimizer_kwargs or {"q": 1, "num_restarts": 10, "raw_samples": 50}
         self.n_candidates = min(5000, max(2000, 200 * problem.dim))
         self.state = BaxusState(dim=problem.dim, eval_budget=self.evaluation_budget - self.n_init)
         self.S = embedding_matrix(input_dim=self.state.dim, target_dim=self.state.d_init)
@@ -131,61 +120,6 @@ class BAxUS(BoTorchAlgorithm):
         if self._acquisition == ExpectedImprovement:
             return self._generate_candidate_qei(tr_lb, tr_ub)
         raise NotImplementedError()
-
-    def _generate_candidate_ts(self, train_X, tr_lb, tr_ub, x_center):
-        dtype = train_X.dtype
-        device = train_X.device
-
-        dim = train_X.shape[-1]
-        sobol = SobolEngine(dim, scramble=True)
-        pert = sobol.draw(self.n_candidates).to(dtype=dtype, device=device)
-        pert = tr_lb + (tr_ub - tr_lb) * pert
-
-        # Create a perturbation mask
-        prob_perturb = min(20.0 / dim, 1.0)
-        mask = torch.rand(self.n_candidates, dim, dtype=dtype, device=device) <= prob_perturb
-        ind = torch.where(mask.sum(dim=1) == 0)[0]
-        mask[ind, torch.randint(0, dim, size=(len(ind),), device=device)] = 1
-
-        # Create candidate points from the perturbations and the mask
-        X_cand = x_center.expand(self.n_candidates, dim).clone()
-        X_cand[mask] = pert[mask]
-
-        # Sample on the candidate points
-        thompson_sampling = self._acquisition(model=self.model, replacement=False)
-        with torch.no_grad():  # We don't need gradients when using TS
-            candidate = thompson_sampling(X_cand, num_samples=1)
-        return candidate
-
-    def _generate_candidate_qei(self, tr_lb, tr_ub):
-        ei = self._acquisition(self.model, self.train_y.max())
-        candidate, _ = optimize_acqf(
-            ei,
-            bounds=torch.stack([tr_lb, tr_ub]),
-            q=1,
-            **self._acquisition_optimizer_kwargs
-        )
-        return candidate
-
-    def _update_state(self, Y_next):
-        if max(Y_next) > self.state.best_value + 1e-3 * math.fabs(self.state.best_value):
-            self.state.success_counter += 1
-            self.state.failure_counter = 0
-        else:
-            self.state.success_counter = 0
-            self.state.failure_counter += 1
-
-        if self.state.success_counter == self.state.success_tolerance:  # Expand trust region
-            self.state.length = min(2.0 * self.state.length, self.state.length_max)
-            self.state.success_counter = 0
-        elif self.state.failure_counter == self.state.failure_tolerance:  # Shrink trust region
-            self.state.length /= 2.0
-            self.state.failure_counter = 0
-
-        self.state.best_value = max(self.state.best_value, max(Y_next).item())
-        if self.state.length < self.state.length_min:
-            self.state.restart_triggered = True
-        return self.state
 
     def _trigger_restart(self):
         self.state.restart_triggered = False
